pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "uk-health-platform"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary",
    "python-dotenv",
    "pyyaml",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]